WIP - fix types and add getProfile to api
